apps/dal-agent/apps/azure-functions-scout/requirements.txt
azure-functions==1.20.0
pyodbc==5.1.0
